Update ds9norm.py
Save the runtime setting of normalization methods and parameters. Users who repeatedly opening the image viewer are free from adjusting the normalization every time when opening. This is convenient for e.g. those who are working on a set of FITS files obtained from the same telescope and hope to use the same normalization parameter for each of those files.